requirements.txt
# FocusBeast v5 — Python dependencies
# Install with: pip install -r requirements.txt

discord.py>=2.3.0
aiosqlite>=0.19.0
Pillow>=10.0.0
